+++ src/test/java/com/nexmo/client/account/AccountClientTest.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2011-2017 Nexmo Inc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.nexmo.client.account;
+
+import junit.framework.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.mockito.Mockito.*;
+
+public class AccountClientTest {
+    private AccountClient client;
+    private BalanceResponse sentinel;
+
+    @Before
+    public void setUp() throws Exception {
+        client = new AccountClient(null);
+        client.balance = mock(BalanceEndpoint.class);
+        sentinel = new BalanceResponse(1.1, true);
+        when(client.balance.execute()).thenReturn(sentinel);
+    }
+
+    @Test
+    public void testGetBalance() throws Exception {
+        BalanceResponse response = client.getBalance();
+        verify(client.balance).execute();
+        Assert.assertEquals(sentinel, response);
+    }
+
+}